+++ migrations/versions/2e2e17b4cdd2_create_currency_table.py
@@ -1,27 +1,27 @@
-"""create levels table
+"""create currency table
 
-Revision ID: 66f698bbb3a5
-Revises:
-Create Date: 2022-12-23 19:11:28.997018
+Revision ID: 2e2e17b4cdd2
+Revises: 66f698bbb3a5
+Create Date: 2022-12-26 23:14:01.120365
 """
 
 from alembic.op import create_table, drop_table
 from sqlalchemy import BigInteger, Column
 
 # revision identifiers, used by Alembic.
-revision = "66f698bbb3a5"
-down_revision = None
+revision = "2e2e17b4cdd2"
+down_revision = "66f698bbb3a5"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     create_table(
-        "levels",
+        "currency",
         Column("id", BigInteger, primary_key=True),
-        Column("experience", BigInteger, default=0, nullable=False),
+        Column("balance", BigInteger, default=0, nullable=False),
     )
 
 
 def downgrade():
-    drop_table("levels")
+    drop_table("currency")